requirements.txt
numpy>=1.24.4
pillow>=9.4.0
tqdm>=4.66.1
hydra-core>=1.3.2
iopath>=0.1.10
einops>=0.8.1